requirements.txt
google-genai==1.72.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
joblib==1.4.2
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
python-dateutil==2.9.0.post0
PyJWT==2.12.1
razorpay==2.0.1
python-decouple==3.8
pytz==2025.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
